[refactor] relocate the round trip tests to converter spec
