+++ packages/lwdita-xdita/test/converter.spec.ts
@@ -17,8 +17,11 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
 import { expect } from 'chai';
-import { jditaToAst, serializeToXdita, xditaToAst, xditaToJdita } from '../src/converter';
+import { astToJdita, jditaToAst, serializeToXdita, xditaToAst, xditaToJdita } from '../src/converter';
 import { DocumentNode } from "@evolvedbinary/lwdita-ast";
+import { fullAstObject, fullJditaObject, fullXditaExample } from './test-utils';
+import { InMemoryTextSimpleOutputStreamCollector } from '../src/stream';
+import { XditaSerializer } from '../src/xdita-serializer';
 
 describe('xditaToAst', () => {
 
@@ -258,4 +261,60 @@ describe('jditaToAst', () => {
     expect(textNode.static.nodeName).to.equal('cdata');
     expect(textNode.readProp("content")).to.equal('cdata');
   })
+});
+
+
+describe('round trip', () => {
+  /**
+   * xdita -> ast -> jdita -> ast -> xdita
+   */
+
+  it('xdita to ast', async () => {
+    const xdita = fullXditaExample;
+    const ast = await xditaToAst(xdita);
+    // assert that the ast is as expected
+    expect(ast).to.deep.equal(fullAstObject);
+  })
+
+  it('ast to jdita', async () => {
+    const xdita = fullXditaExample;
+    const ast = await xditaToAst(xdita);
+
+    const jdita = astToJdita(ast);
+    // assert that the lwdita is as expected
+    expect(jdita).to.deep.equal(fullJditaObject);
+  })
+
+  it('jdita to ast', async () => {
+    const xdita = fullXditaExample;
+    const ast = await xditaToAst(xdita);
+    // assert that the ast is as expected
+    expect(ast).to.deep.equal(fullAstObject);
+
+    const jdita = astToJdita(ast);
+
+    const newAst = jditaToAst(jdita);
+    // assert that the new ast is as expected
+    expect(newAst).to.deep.equal(fullAstObject);
+  });
+
+  it('ast to xdita', async () => {
+    const xdita = fullXditaExample;
+    const ast = await xditaToAst(xdita);
+
+    const jdita = astToJdita(ast);
+
+    const newAst = jditaToAst(jdita);
+
+    const outStream = new InMemoryTextSimpleOutputStreamCollector();
+    const serializer = new XditaSerializer(outStream);
+
+    serializer.serialize(newAst);
+
+    const newXdita = outStream.getText();
+
+    // assert that the new xdita is as expected
+    expect(newXdita).to.equal(xdita);
+  });
+
 });
